feat: implement authentication system with JWT and email verification

+++ splendid-backend/src/main/java/track/expense/splendid_backend/service/impl/TransactionServiceImpl.java
@@ -1,12 +1,15 @@
 package track.expense.splendid_backend.service.impl;
 
 import lombok.AllArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import track.expense.splendid_backend.dto.TransactionDto;
 import track.expense.splendid_backend.entity.Transaction;
+import track.expense.splendid_backend.entity.User;
 import track.expense.splendid_backend.exception.ResourceNotFoundException;
 import track.expense.splendid_backend.mapper.TransactionMapper;
 import track.expense.splendid_backend.repository.TransactionRepository;
+import track.expense.splendid_backend.repository.UserRepository;
 import track.expense.splendid_backend.service.TransactionService;
 
 import java.util.List;
@@ -17,12 +20,17 @@ import java.util.stream.Collectors;
 public class TransactionServiceImpl implements TransactionService {
 
     private TransactionRepository transactionRepository;
+    private UserRepository userRepository;
 
     @Override
     public TransactionDto createTransaction(TransactionDto transactionDto) {
         Transaction transaction = TransactionMapper.toEntity(transactionDto);
-        Transaction savedTransaction = transactionRepository.save(transaction);
 
+        String email = SecurityContextHolder.getContext().getAuthentication().getName();
+        User user = userRepository.findByEmail(email).orElseThrow(() -> new ResourceNotFoundException("User not found"));
+        transaction.setUser(user);
+
+        Transaction savedTransaction = transactionRepository.save(transaction);
         return TransactionMapper.toDto(savedTransaction);
     }
 
@@ -34,13 +42,21 @@ public class TransactionServiceImpl implements TransactionService {
 
     @Override
     public List<TransactionDto> getAllTransactions() {
-        List<Transaction> transactions = transactionRepository.findAll();
+        String email = SecurityContextHolder.getContext().getAuthentication().getName();
+        User user = userRepository.findByEmail(email).orElseThrow(() -> new ResourceNotFoundException("User not found"));
+        List<Transaction> transactions = transactionRepository.findByUser(user);
         return transactions.stream().map(TransactionMapper::toDto).toList();
     }
 
     @Override
     public TransactionDto updateTransaction(Long transactionId, TransactionDto transactionDto) {
         Transaction transaction = transactionRepository.findById(transactionId).orElseThrow(() -> new ResourceNotFoundException("Transaction not found with this id : " + transactionId));
+        String email = SecurityContextHolder.getContext().getAuthentication().getName();
+
+        if (!transaction.getUser().getEmail().equals(email)) {
+            throw new ResourceNotFoundException("unauthorized access");
+        }
+
         transaction.setTitle(transactionDto.getTitle());
         transaction.setAmount(transactionDto.getAmount());
         transaction.setType(Transaction.TransactionType.valueOf(transactionDto.getType().toUpperCase()));
@@ -55,7 +71,11 @@ public class TransactionServiceImpl implements TransactionService {
 
     @Override
     public void deleteTransaction(Long transactionId) {
-        transactionRepository.findById(transactionId).orElseThrow(() -> new ResourceNotFoundException("Transaction not found with this id : " + transactionId));
+        Transaction transaction = transactionRepository.findById(transactionId).orElseThrow(() -> new ResourceNotFoundException("Transaction not found with this id : " + transactionId));
+        String email = SecurityContextHolder.getContext().getAuthentication().getName();
+        if (!transaction.getUser().getEmail().equals(email)) {
+            throw new ResourceNotFoundException("unauthorized access");
+        }
         transactionRepository.deleteById(transactionId);
     }
